Move project cloning steps to Getting Started page
The steps for cloning the project to a local machine apply to all usage paths (i.e. development container, command line, and UI deployments), but are only documented in this development container page. The "setup" steps in this page (e.g. the steps of getting the bits to the local machine and ready to run) should be moved to top-level "getting started" page. This will also help simplify/clarify the "Command-Line Deployment" steps by avoiding some context switching across the different getting started pages.