Update tier3.bicep Adding Defender configuration to Tier 3 (#725)
* Update tier3.bicep

* Tier 3 Defender Update

Pulling Defender info from the core mlz.bicep deployment
Updating Tier 3 Readme

* Update Tier3. Removing Duplicate parameters and moving comments up

Fixing error with duplicate parameters. Adding Moving description comments to the parameters that pull from mlzDeploymentVariables

* Adding New Outputs to the Bicep Core Readme Documentation

Adding New Outputs to the Bicep Core Readme Documentation

* Removing parameters from the Tier3 Readme that are not needed because they are inherited

With the inheritance pattern proposed in this PR, this will not be necessary to explicitly state here. The state of just using the same default configuration in the original MLZ deployment will be sufficient with the added changes of documenting the additional two outputs.

Co-authored-by: lisamurphy-msft <87032539+lisamurphy-msft@users.noreply.github.com>